backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.25
pydantic==2.6.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
python-dotenv==1.0.1
boto3==1.34.34
pillow==10.2.0
opencv-python==4.9.0.80
torch==2.2.0
torchvision==0.17.0
transformers==4.37.2
diffusers==0.26.3
pytest==8.0.0
httpx==0.26.0
alembic==1.13.1 